+++ lib/flutter_joystick.dart
@@ -2,6 +2,7 @@ library flutter_joystick;
 
 export 'src/joystick.dart';
 export 'src/joystick_area.dart';
+export 'src/joystick_arrow.dart';
 export 'src/joystick_base.dart';
 export 'src/joystick_controller.dart';
 export 'src/joystick_stick.dart';