+++ lib/src/joystick_base.dart
@@ -1,51 +1,32 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_joystick/src/joystick_arrow.dart';
-import 'joystick.dart';
+import 'package:flutter_joystick/flutter_joystick.dart';
 
 class JoystickBase extends StatelessWidget {
-  final Color? arrowsColor;
-  final List<BoxShadow>? boxShadows;
-  final Color color;
-  final bool drawArrows;
   final JoystickMode mode;
   final double size;
-  final bool withBorderCircle;
-  final bool enableArrowAnimation;
+  final JoystickBaseDecoration? decoration;
+  final Widget? joystickArrows;
+  final JoystickArrowsDecoration? arrowsDecoration;
 
   const JoystickBase({
-    this.arrowsColor,
-    this.boxShadows,
-    this.color = const Color(0x50616161),
-    this.drawArrows = true,
     this.mode = JoystickMode.all,
     this.size = 200,
-    this.withBorderCircle = true,
-    this.enableArrowAnimation = true,
+    this.decoration,
+    this.joystickArrows,
+    this.arrowsDecoration,
     super.key,
-  });
+  }) : assert(joystickArrows == null || arrowsDecoration == null);
 
   @override
   Widget build(BuildContext context) {
-    Color boxShadowColor =
-        MediaQuery.of(context).platformBrightness == Brightness.dark
-            ? color.withOpacity(0.08)
-            : color.withOpacity(0.16);
+    final baseDecoration = decoration ?? JoystickBaseDecoration();
     return Container(
       width: size,
       height: size,
       decoration: BoxDecoration(
-        color: Colors.transparent,
-        shape: BoxShape.circle,
-        boxShadow: boxShadows ??
-            [
-              BoxShadow(
-                color: boxShadowColor,
-                spreadRadius: 1,
-                blurRadius: 12,
-                offset: const Offset(0, 4),
-              ),
-            ],
-      ),
+          color: Colors.transparent,
+          shape: BoxShape.circle,
+          boxShadow: baseDecoration.boxShadows),
       child: Stack(
         children: [
           SizedBox(
@@ -53,19 +34,17 @@ class JoystickBase extends StatelessWidget {
             height: size,
             child: CustomPaint(
               painter: _JoystickBasePainter(
-                color: color,
-                mode: mode,
-                withBorderCircle: withBorderCircle,
+                decoration: baseDecoration,
               ),
             ),
           ),
-          if (drawArrows)
-            JoystickArrows(
-              arrowsColor: arrowsColor ?? color,
-              mode: mode,
-              size: size,
-              enableAnimation: enableArrowAnimation,
-            ),
+          if (baseDecoration.drawArrows)
+            joystickArrows ??
+                JoystickArrows(
+                  mode: mode,
+                  size: size,
+                  decoration: arrowsDecoration,
+                ),
         ],
       ),
     );
@@ -73,14 +52,10 @@ class JoystickBase extends StatelessWidget {
 }
 
 class _JoystickBasePainter extends CustomPainter {
-  final Color color;
-  final JoystickMode mode;
-  final bool withBorderCircle;
+  final JoystickBaseDecoration decoration;
 
   _JoystickBasePainter({
-    required this.color,
-    required this.mode,
-    required this.withBorderCircle,
+    required this.decoration,
   });
 
   static const double borderStrokeWidthPercentage = 0.05;
@@ -93,37 +68,47 @@ class _JoystickBasePainter extends CustomPainter {
     final radius = diameter / 2;
     final center = Offset(radius, radius);
 
-    drawCircles(canvas, center, diameter, withBorderCircle);
+    drawCircles(canvas, center, diameter);
   }
 
-  void drawCircles(
-      Canvas canvas, Offset center, double diameter, bool withBorderCircle) {
-    if (withBorderCircle) {
-      final borderPaint = Paint()
-        ..color = color
+  void drawCircles(Canvas canvas, Offset center, double diameter) {
+    drawOuterCircle(canvas, center, diameter);
+    drawMiddleCircle(canvas, center, diameter);
+    drawInnerCircle(canvas, center, diameter);
+  }
+
+  void drawInnerCircle(Canvas canvas, Offset center, double diameter) {
+    if (decoration.drawInnerCircle) {
+      final innerPaint = Paint()
+        ..color = decoration.innerCircleColor
+        ..style = PaintingStyle.fill;
+      canvas.drawCircle(
+          center,
+          diameter / 2 - outermostCircleRadiusReductionPercentage * diameter,
+          innerPaint);
+    }
+  }
+
+  void drawMiddleCircle(Canvas canvas, Offset center, double diameter) {
+    if (decoration.drawMiddleCircle) {
+      final paint = Paint()
+        ..color = decoration.middleCircleColor
+        ..style = PaintingStyle.fill;
+      canvas.drawCircle(
+          center,
+          diameter / 2 - innerCircleRadiusReductionPercentage * diameter,
+          paint);
+    }
+  }
+
+  void drawOuterCircle(Canvas canvas, Offset center, double diameter) {
+    if (decoration.drawOuterCircle) {
+      final outerCirclePaint = Paint()
+        ..color = decoration.outerCircleColor
         ..strokeWidth = borderStrokeWidthPercentage * diameter
         ..style = PaintingStyle.stroke;
-      canvas.drawCircle(center, diameter / 2, borderPaint);
+      canvas.drawCircle(center, diameter / 2, outerCirclePaint);
     }
-
-    final outerPaint = Paint()
-      ..color = color.withOpacity(0.84)
-      ..style = PaintingStyle.fill;
-
-    final innerPaint = Paint()
-      ..color = color
-      ..style = PaintingStyle.fill;
-
-    canvas.drawCircle(
-      center,
-      diameter / 2 - innerCircleRadiusReductionPercentage * diameter,
-      outerPaint,
-    );
-    canvas.drawCircle(
-      center,
-      diameter / 2 - outermostCircleRadiusReductionPercentage * diameter,
-      innerPaint,
-    );
   }
 
   @override
@@ -131,36 +116,39 @@ class _JoystickBasePainter extends CustomPainter {
 }
 
 class JoystickSquareBase extends StatelessWidget {
+  final test = JoystickBaseDecoration();
   final JoystickMode mode;
   final double size;
-  final bool drawArrows;
-  final Color color;
+  final JoystickBaseDecoration? decoration;
 
-  const JoystickSquareBase({
+  JoystickSquareBase({
     this.mode = JoystickMode.all,
     this.size = 200,
-    this.drawArrows = true,
-    this.color = const Color(0x50616161),
+    this.decoration,
     super.key,
   });
 
   @override
   Widget build(BuildContext context) {
+    final decoration = this.decoration ?? JoystickBaseDecoration();
     final padding = 10 / 200 * size;
     return Container(
       width: size,
       height: size,
-      decoration: BoxDecoration(border: Border.all(color: color, width: 10)),
+      decoration: decoration.drawOuterCircle
+          ? BoxDecoration(
+              border: Border.all(color: decoration.outerCircleColor, width: 10))
+          : null,
       child: Padding(
         padding: EdgeInsets.all(padding),
         child: Container(
           width: size - padding * 2,
           height: size - padding * 2,
-          color: color,
-          child: drawArrows
+          color: decoration.middleCircleColor,
+          child: decoration.drawArrows
               ? CustomPaint(
-                  painter:
-                      _JoystickSquareBaseArrowPainter(mode: mode, color: color),
+                  painter: _JoystickSquareBaseArrowPainter(
+                      mode: mode, decoration: decoration),
                 )
               : null,
         ),
@@ -171,11 +159,11 @@ class JoystickSquareBase extends StatelessWidget {
 
 class _JoystickSquareBaseArrowPainter extends CustomPainter {
   final JoystickMode mode;
-  final Color color;
+  final JoystickBaseDecoration decoration;
 
   _JoystickSquareBaseArrowPainter({
     required this.mode,
-    required this.color,
+    required this.decoration,
   });
 
   @override
@@ -185,7 +173,7 @@ class _JoystickSquareBaseArrowPainter extends CustomPainter {
     final linePosition = 30.0 / 180 * size.width;
     final double arrowSpacing = 15.0 / 180 * size.width;
     final linePaint = Paint()
-      ..color = color
+      ..color = decoration.innerCircleColor
       ..strokeWidth = 5 / 180 * size.width
       ..style = PaintingStyle.stroke;
 
@@ -263,4 +251,67 @@ class _JoystickSquareBaseArrowPainter extends CustomPainter {
 
   @override
   bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
+}
+
+@immutable
+class JoystickBaseDecoration {
+  final bool drawArrows;
+
+  final bool drawOuterCircle;
+  final Color outerCircleColor;
+  final bool drawMiddleCircle;
+  final Color middleCircleColor;
+  final bool drawInnerCircle;
+  final Color innerCircleColor;
+
+  final List<BoxShadow>? boxShadows;
+
+  const JoystickBaseDecoration._internal({
+    required this.drawArrows,
+    required this.drawOuterCircle,
+    required this.outerCircleColor,
+    required this.drawMiddleCircle,
+    required this.middleCircleColor,
+    required this.drawInnerCircle,
+    required this.innerCircleColor,
+    required this.boxShadows,
+  });
+
+  factory JoystickBaseDecoration({
+    Color? color,
+    bool drawArrows = true,
+    bool drawOuterCircle = true,
+    Color? outerCircleColor,
+    bool drawMiddleCircle = true,
+    Color? middleCircleColor,
+    bool drawInnerCircle = true,
+    Color? innerCircleColor,
+    List<BoxShadow>? boxShadows,
+    Color? boxShadowColor,
+  }) {
+    assert(boxShadows == null || boxShadowColor == null);
+    assert(color == null ||
+        outerCircleColor == null ||
+        middleCircleColor == null ||
+        innerCircleColor == null);
+    Color baseColor = color ?? ColorUtils.defaultBaseColor;
+    return JoystickBaseDecoration._internal(
+      drawArrows: drawArrows,
+      drawOuterCircle: drawOuterCircle,
+      outerCircleColor: outerCircleColor ?? baseColor,
+      drawMiddleCircle: drawMiddleCircle,
+      middleCircleColor: middleCircleColor ?? ColorUtils.lighten(baseColor, 0.05),
+      drawInnerCircle: drawInnerCircle,
+      innerCircleColor: innerCircleColor ?? baseColor,
+      boxShadows: boxShadows ??
+          [
+            BoxShadow(
+              color: boxShadowColor ?? baseColor.withOpacity(0.16),
+              spreadRadius: 10,
+              blurRadius: 12,
+              offset: const Offset(0, 4),
+            ),
+          ],
+    );
+  }
 }