+++ lib/src/joystick_stick.dart
@@ -2,19 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_joystick/src/utils.dart';
 
 class JoystickStick extends StatelessWidget {
-  final Color color;
-  final Color? shadowColor;
   final double size;
+  final JoystickStickDecoration? decoration;
 
   const JoystickStick({
-    this.color = Colors.lightBlue,
-    this.shadowColor,
     this.size = 50,
+    this.decoration,
     super.key,
   });
 
   @override
   Widget build(BuildContext context) {
+    final decoration = this.decoration ?? JoystickStickDecoration();
     return Container(
       width: size,
       height: size,
@@ -22,7 +21,7 @@ class JoystickStick extends StatelessWidget {
         shape: BoxShape.circle,
         boxShadow: [
           BoxShadow(
-            color: shadowColor ?? color.withOpacity(0.5),
+            color: decoration.shadowColor,
             spreadRadius: 5,
             blurRadius: 7,
             offset: const Offset(0, 3),
@@ -32,11 +31,32 @@ class JoystickStick extends StatelessWidget {
           begin: Alignment.topCenter,
           end: Alignment.bottomCenter,
           colors: [
-            ColorUtils.darken(color),
-            ColorUtils.lighten(color),
+            ColorUtils.darken(decoration.color),
+            ColorUtils.lighten(decoration.color),
           ],
         ),
       ),
+    );
+  }
+}
+
+@immutable
+class JoystickStickDecoration {
+  final Color color;
+  final Color shadowColor;
+
+  const JoystickStickDecoration._internal({
+    required this.color,
+    required this.shadowColor,
+  });
+
+  factory JoystickStickDecoration({
+    Color color = ColorUtils.defaultStickColor,
+    Color? shadowColor,
+  }) {
+    return JoystickStickDecoration._internal(
+      color: color,
+      shadowColor: shadowColor ?? color.withOpacity(0.5),
     );
   }
 }